Add more abstract functions.  Modified the function CreateAssemblyViaInputFile to be more generic.
