+++ src/CADAssembler/CADCommonFunctions/CADCommonFunctions.vcxproj
@@ -164,8 +164,10 @@
   <ItemGroup>
     <ClInclude Include="..\..\..\generated\CAD\generated\AssemblyInterface.h" />
     <ClInclude Include="..\..\..\generated\CAD\generated\AssemblyInterface_xsd.h" />
+    <ClInclude Include="cc_AssemblyUtilities.h" />
     <ClInclude Include="c3ga.h" />
     <ClInclude Include="cc_AssemblyOptions.h" />
+    <ClInclude Include="cc_CADFactoryAbstract.h" />
     <ClInclude Include="cc_CommonConstants.h" />
     <ClInclude Include="cc_CommonDefinitions.h" />
     <ClInclude Include="cc_CommonFunctions.h" />
@@ -192,6 +194,7 @@
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="..\..\..\generated\CAD\generated\AssemblyInterface.cpp" />
+    <ClCompile Include="cc_AssemblyUtilities.cpp" />
     <ClCompile Include="c3ga.cpp" />
     <ClCompile Include="cc_AssemblyOptions.cpp" />
     <ClCompile Include="cc_CommonDefinitions.cpp" />
