+++ src/CADAssembler/CADCommonFunctions/cc_LoggerBoost.h
@@ -58,10 +58,10 @@ void SetupLogging(	const std::string		&in_SubDir,
 
 void	LogMainNonZeroExitCode( const std::string			&in_ExeName,
 								int							in_ExitCode, 
-								const std::stringstream	&in_InputLine,
+								const std::string			&in_InputLine,
 								bool						in_Logging_Set_Up, 
 								const std::string			in_LogFileName,  
-								const std::stringstream		&in_ExceptionErrorStringStream );
+								const std::string			&in_ExceptionErrorStringStream );
 
 
 namespace isis {
